requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
filelock==3.20.0
fonttools==4.54.1
fsspec==2024.9.0
idna==3.11
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==3.0.3
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==2.1.3
opencv-python==4.12.0.88
packaging==24.1
pillow==10.4.0
polars==1.5.0
psutil==6.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==80.9.0
six==1.17.0
supervision==0.26.1
threadpoolctl==3.5.0
torch==2.9.0
tqdm==4.66.5
typing_extensions==4.12.2
ultralytics==8.3.221
ultralytics-thop==2.0.17
urllib3==2.2.3
